[HAMA-575]: Generify graph package
git-svn-id: https://svn.apache.org/repos/asf/incubator/hama/trunk@1340131 13f79535-47bb-0310-9956-ffa450edef68

+++ examples/src/main/java/org/apache/hama/examples/CombineExample.java
@@ -118,7 +118,7 @@ public class CombineExample {
     if (bsp.waitForCompletion(true)) {
       printOutput(conf);
       System.out.println("Job Finished in "
-          + (double) (System.currentTimeMillis() - startTime) / 1000.0
+          + (System.currentTimeMillis() - startTime) / 1000.0
           + " seconds");
     }
